requirements.txt
# Core dependencies for RAG agent
langchain>=0.1.0
langchain-community>=0.0.10

# PDF processing
PyPDF2>=3.0.0
pypdf>=3.17.0

# Vector database
faiss-cpu>=1.7.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests for NVIDIA API
requests>=2.31.0

# Additional utilities
numpy>=1.21.0
tqdm>=4.65.0

# Web interface
streamlit>=1.28.0
plotly>=5.15.0

# Web interface
streamlit>=1.28.0
plotly>=5.15.0
